All tests pass save 1 & you're not gonna like why.
I think this needs to record the expression tree so that the tree can be walked freely instead of only forward. I think this is causing problems with OR expectations, b/c failed children of an OR expectation can't roll back walking of the expression tree.

I think the condition chain concept may also need to move to a tree.